requirements.txt
chainlit>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0
google-genai>=1.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
pdfplumber>=0.10.0
python-docx>=1.1.0
python-pptx>=0.6.23
psutil>=5.9.0
